requirements.txt
pymetasploit3
termcolor
python-nmap
requests
flask
tensorflow==2.7.3
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
matplotlib==3.4.3
pandas==1.1.5
numpy==1.16.3
numpydoc==0.9.1
